finish

+++ src/main/java/fr/esgi/extiaordinaryapi/service/SeanceService.java
@@ -1,7 +1,9 @@
 package fr.esgi.extiaordinaryapi.service;
 
 import fr.esgi.extiaordinaryapi.entity.Seance;
+import fr.esgi.extiaordinaryapi.entity.User;
 import fr.esgi.extiaordinaryapi.repository.SeanceRepository;
+import fr.esgi.extiaordinaryapi.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
@@ -19,6 +21,8 @@ public class SeanceService {
     private final int MAX_POINT = 445;
 
     private final SeanceRepository seanceRepository;
+
+    private final UserRepository userRepository;
 
     public Seance createSeance(Seance seance) {
         try {
@@ -38,13 +42,20 @@ public class SeanceService {
         }
     }
 
-    public Seance updateSeance(Seance seance) {
+    public Seance updateSeance(Seance seance, User user) {
         if (seance.getRewardPoint() > MAX_POINT) {
             throw new IllegalArgumentException("Reward point is too high");
         }
         val findSeance = seanceRepository.findById(seance.getSeanceId());
         if (findSeance.isEmpty()) {
             throw new IllegalArgumentException("Seance not found");
+        }
+        if(findSeance.get().getCoach().getId() != user.getId()) {
+            throw new IllegalArgumentException("User is not coach");
+        }
+        val localDateTime = LocalDateTime.now();
+        if (findSeance.get().getDateStart().isBefore(localDateTime)) {
+            throw new IllegalArgumentException("Seance is already started");
         }
         return createSeance(
                 Seance.builder()
@@ -55,6 +66,7 @@ public class SeanceService {
                         .dateEnd(seance.getDateEnd())
                         .rewardPoint(seance.getRewardPoint())
                         .creationDate(findSeance.get().getCreationDate())
+                        .coach(seance.getCoach())
                         .build()
         );
     }
@@ -71,16 +83,56 @@ public class SeanceService {
         return seanceRepository.findAll();
     }
 
-    public void deleteSeance(UUID id) {
+    public void deleteSeance(UUID id, User user) {
         val findSeance = seanceRepository.findById(id);
         if (findSeance.isEmpty()) {
             throw new IllegalArgumentException("Seance not found");
+        }
+        if (findSeance.get().getCoach().getId() != user.getId()) {
+            throw new IllegalArgumentException("User is not coach");
+        }
+        val localDateTime = LocalDateTime.now();
+        if (findSeance.get().getDateStart().isBefore(localDateTime)) {
+            throw new IllegalArgumentException("Seance is already started");
         }
         try {
             seanceRepository.delete(findSeance.get());
         } catch (Exception e) {
             log.error("Error while deleting seance", e);
             throw new RuntimeException("Error while deleting seance", e);
+        }
+    }
+
+    public String addUserToSeance(UUID idSeance, User user) {
+        val findSeance = seanceRepository.findById(idSeance);
+        if (findSeance.isEmpty()) {
+            throw new IllegalArgumentException("Seance not found");
+        }
+        if (findSeance.get().getCoach().getId() == user.getId()) {
+            throw new IllegalArgumentException("User is coach");
+        }
+        val localDateTime = LocalDateTime.now();
+        if (findSeance.get().getDateStart().isBefore(localDateTime)) {
+            throw new IllegalArgumentException("Seance is already started");
+        }
+        try {
+            val addSeanceToUser = user.getSeancesPlayed();
+            addSeanceToUser.add(findSeance.get());
+            userRepository.save(User.builder()
+                    .id(user.getId())
+                    .firstName(user.getFirstName())
+                    .lastName(user.getLastName())
+                    .email(user.getEmail())
+                    .password(user.getPassword())
+                    .dateOfBirth(user.getDateOfBirth())
+                    .points(user.getPoints())
+                    .role(user.getRole())
+                    .seancesPlayed(addSeanceToUser)
+                    .build());
+            return "User added to seance";
+        } catch (Exception e) {
+            log.error("Error while adding seance to user", e);
+            throw new RuntimeException("Error while adding seance to user", e);
         }
     }
 }
